LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/last-successful-intermediate-ignored-expected.txt


PASS Starts rendering with flip-block
FAIL No successful position (with intermediate successful), keep flip-block assert_equals: expected 200 but got 0
